// collection-service/src/adapters/module-exchange-client/types.ts
export type BalancesResponse<R = any> = {
  info: R; // the original untouched non-parsed reply with details
  timestamp: number; // Unix Timestamp in milliseconds (seconds * 1000)
  datetime: string; // ISO8601 datetime string with milliseconds

  //-------------------------------------------------------------------------
  // indexed by availability of funds first, then by currency

  data: {
    [accountType: string]: {
      [code: string]: {
        id: string;
        code: string;
        free: string;
        used: string; // money on hold, locked, frozen, or pending, by currency
        total: string; // total (free + used), by currency
      };
    };
  };

  //-------------------------------------------------------------------------
  // indexed by currency first, then by availability of funds
};